+++ eob2/answers.h
@@ -23,9 +23,9 @@ struct answers {
 	void			clear();
 	static int		compare(const void* v1, const void* v2);
 	const element*	end() const { return elements.end(); }
-	int				findvalue(const void* pv) const;
+	size_t			findvalue(const void* pv) const;
 	bool			have(const void* pv) const { return elements.have(pv); }
-	int				getcount() const { return elements.getcount(); }
+	size_t			getcount() const { return elements.getcount(); }
 	const char*		getname(void* v);
 	int				indexof(const void* v) const { return elements.indexof(v); }
 	void*			random() const;
