+++ eob2/slice.h
@@ -16,8 +16,8 @@ extern "C" void* memchr(const void* ptr, int value, long unsigned num);
 extern "C" void* memcpy(void* destination, const void* source, long unsigned size) noexcept(true);
 extern "C" int memcmp(const void* p1, const void* p2, size_t size) noexcept(true);
 extern "C" void* memmove(void* destination, const void* source, size_t size) noexcept(true);
-extern "C" void* memset(void* destination, int value, long unsigned size) noexcept(true);
-extern "C" void	qsort(void* base, unsigned num, long unsigned size, fncompare proc);
+extern "C" void* memset(void* destination, int value, size_t size) noexcept(true);
+extern "C" void	qsort(void* base, size_t num, size_t size, fncompare proc);
 
 template<class T> class slice {
 	T* data;
