+++ eob2/stringbuilder.h
@@ -59,7 +59,7 @@ public:
 	const char*		end() const { return pe; }
 	char*			get() const { return p; }
 	static int		getgender(const char* s);
-	unsigned		getmaximum() const { return pe - pb - 1; }
+	size_t			getmaximum() const { return pe - pb - 1; }
 	bool			isempthy() const { return !pb || pb[0] == 0; }
 	bool			isfull() const { return p >= pe; }
 	bool			ispos(const char* v) const { return p == v; }
@@ -69,7 +69,7 @@ public:
 	const char*	psstr(const char* p, char end_symbol);
 	const char*	psstrlf(const char* p);
 	void			set(char* v) { p = v; p[0] = 0; }
-	unsigned		size() const { return p - pb; }
+	size_t			size() const { return p - pb; }
 	void			trimr();
 	void			upper();
 };
@@ -118,7 +118,7 @@ constexpr bool ischa(unsigned char s) { return (s >= 'A' && s <= 'Z') || (s >= '
 constexpr bool isnum(unsigned char s) { return s >= '0' && s <= '9'; }
 constexpr const char* zchr(const char* p, char e) { while(*p) { if(*p == e) return p; p++; } return 0; }
 constexpr const char* zend(const char* p) { while(*p) p++; return p; }
-constexpr unsigned long zlen(const char* p) { return zend(p) - p; }
+constexpr size_t zlen(const char* p) { return zend(p) - p; }
 
 template<class T> void ftinfo(const void* object, stringbuilder& sb);
